In memory model transfer between neuron and coreneuron (#114)
This is the first of a series of transformations to allow CoreNeuron to be embedded in NEURON and simulate the NEURON model without using data files. This change reads the bbcore_mech.dat file as an ifstream or can callback, if embedded, to get the bbcore_mech.dat file information directly from NEURON via a stringstream. In both cases the information is handled by the same mk_mech(std::istream&) function.

- binary sentinel value for determining Endian consistency has been removed from bbcore_mech.dat and into a binary file called byteswap1.dat
- bbcore_write_version has been bumped to 1.2
- direct transfer of globals.dat info in place
- update test dataset
- if OMP_NUM_THREADS is not set then it is set to nthread.